+++ test/pokemon/get-pokemon.spec.js
@@ -21,9 +21,9 @@ class PokemonImgMock {
 
 // PokemonObjのMockクラス
 class PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) {
+    constructor ({id}) {
         this.id = id;
-        this.img = new PokemonImg(this).getImgPath();
+        this.img = new PokemonImgMock(this).getImgPath();
     }
     getId () { return this.id; }
     getIsShiny () { return false; }
@@ -33,36 +33,36 @@ class PokemonObjMockNormal {
     getImg () { return this.img; }
 }
 class PokemonObjMockShinyGod extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getIsShiny () { return true; }
     getStrengthLv () { return 'god'; }
 }
 class PokemonObjMockShiny extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getIsShiny () { return true; }
 }
 class PokemonObjMockGod extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getStrengthLv () { return 'god'; }
 }
 class PokemonObjMockStrongest extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getStrengthLv () { return 'strongest'; }
 }
 class PokemonObjMockStronger extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getStrengthLv () { return 'stronger'; }
 }
 class PokemonObjMockWeaker extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getStrengthLv () { return 'weaker'; }
 }
 class PokemonObjMockWeakest extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getStrengthLv () { return 'weakest'; }
 }
 class PokemonObjMockOther extends PokemonObjMockNormal {
-    constructor ({id}, PokemonImg) { super({id}, PokemonImg); }
+    constructor ({id}) { super({id}); }
     getStrengthLv () { return 'hoge'; }
 }
 
@@ -85,7 +85,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('constructorにRequestApiクラスとuserを渡すと正しく処理する', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -99,7 +98,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('constructorにRequestApiクラスとuserを渡さないとErrorをthrowする（RequestApiがnull）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -112,7 +110,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('constructorにRequestApiクラスとuserを渡さないとErrorをthrowする（RequestApiがundefined）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -125,7 +122,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('constructorにRequestApiクラスとuserを渡さないとErrorをthrowする（userがnull）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -138,7 +134,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('constructorにRequestApiクラスとuserを渡さないとErrorをthrowする（userがundefined）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -151,7 +146,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('constructorにRequestApiクラスとuserを渡さないとErrorをthrowする（RequestApiおよびuserが空）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -164,7 +158,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（通常）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -177,7 +170,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（色違いおよびgod）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockShinyGod }
         }).default;
 
@@ -190,7 +182,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（色違い）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockShiny }
         }).default;
 
@@ -203,7 +194,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（god）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockGod }
         }).default;
 
@@ -216,7 +206,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（strongest）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockStrongest }
         }).default;
 
@@ -229,7 +218,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（stronger）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockStronger }
         }).default;
 
@@ -242,7 +230,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（weaker）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockWeaker }
         }).default;
 
@@ -255,7 +242,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（weakest）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockWeakest }
         }).default;
 
@@ -268,7 +254,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッドを呼ぶとランダムにポケモン捕獲テキストを返却する（other）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockOther }
         }).default;
 
@@ -281,7 +266,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッド内でエラーが発生するとErrorをthrowする（request失敗）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -297,7 +281,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('getRandom()メソッド内でエラーが発生するとErrorをthrowする（dataがnull）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -313,7 +296,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('pushData()メソッドを呼ぶと捕まえたポケモンを保存する', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
         // new Date()をMockに差し替え(※タイムゾーン0)
@@ -333,7 +315,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('pushData()メソッド内でエラーが発生するとErrorをthrowする（push失敗）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -348,7 +329,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('pushData()メソッド内でエラーが発生するとErrorをthrowする（引数無し）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -363,7 +343,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('pushData()メソッド内でエラーが発生するとErrorをthrowする（捕獲前）', async () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -377,7 +356,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('テキストを正しく返却する（GO_RES）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -387,7 +365,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('テキストを正しく返却する（GET_ERROR_RES）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -397,7 +374,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('テキストを正しく返却する（PUSH_ERROR_RES）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
 
@@ -407,7 +383,6 @@ describe('pokemon/get-pokemon.js', () => {
     it('テキストを正しく返却する（TIME_FORMAT）', () => {
         // get-pokemon.js内のimportをダミーに差し替え
         const GetPokemon = proxyquire('../../src/pokemon/get-pokemon', {
-            './pkparaiso': { default: PokemonImgMock },
             './pokemon-ja': { default: PokemonObjMockNormal }
         }).default;
